fix: BE api 요청시 쿠키값 확인 후 해당 유저 ACTIVE 상태시 처리가능

+++ src/main/java/com/nova/narrativa/domain/user/controller/S3ImgController.java
@@ -2,6 +2,7 @@ package com.nova.narrativa.domain.user.controller;
 
 
 import com.nova.narrativa.domain.user.service.S3ImageService;
+import com.nova.narrativa.domain.user.service.SignUpService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -19,20 +20,26 @@ import java.util.Map;
 public class S3ImgController {
 
     private final S3ImageService s3ImageService;
+    private final SignUpService signUpService;
 
     // 이미지 업로드
     @PostMapping("/images/upload")
-    public ResponseEntity<?> s3Upload(@RequestPart(value = "image", required = false) MultipartFile image) {
+    public ResponseEntity<?> s3Upload(@RequestPart(value = "image", required = false) MultipartFile image,
+                                      @CookieValue Long id) {
         if (image == null || image.isEmpty()) {
             log.error("이미지 파일이 비어 있습니다.");
             return ResponseEntity.badRequest().body("이미지 파일이 비어 있습니다.");
         }
 
         try {
-            String uploadedImageUrl = s3ImageService.upload(image);
-            Map<String, String> response = new HashMap<>();
-            response.put("imageUrl", uploadedImageUrl);
-            return ResponseEntity.ok(response);
+            if (signUpService.isUserActive(id)) {   // ACTIVE 유저 인 경우만
+                String uploadedImageUrl = s3ImageService.upload(image);
+                Map<String, String> response = new HashMap<>();
+                response.put("imageUrl", uploadedImageUrl);
+                return ResponseEntity.ok(response);
+            } else {
+                return ResponseEntity.notFound().build();
+            }
         } catch (Exception e) {
             log.error("이미지 업로드 중 오류 발생", e);
             return ResponseEntity.status(500).body("이미지 업로드 실패");
@@ -41,10 +48,15 @@ public class S3ImgController {
 
     // S3 버킷에 있는 이미지 삭제
     @DeleteMapping("/images")
-    public ResponseEntity<?> s3Delete(@RequestParam(value = "imageUrl") String imageUrl) {
+    public ResponseEntity<?> s3Delete(@RequestParam(value = "imageUrl") String imageUrl,
+                                      @CookieValue Long id) {
         try {
-            s3ImageService.deleteImageFromS3(imageUrl);
-            return ResponseEntity.ok("이미지 삭제 완료");
+            if (signUpService.isUserActive(id)) {   // ACTIVE 유저 인 경우만
+                s3ImageService.deleteImageFromS3(imageUrl);
+                return ResponseEntity.ok("이미지 삭제 완료");
+            } else {
+                return ResponseEntity.notFound().build();
+            }
         } catch (Exception e) {
             log.error("이미지 삭제 중 오류 발생", e);
             return ResponseEntity.status(500).body("이미지 삭제 실패");
@@ -53,12 +65,17 @@ public class S3ImgController {
 
     // S3 버킷에 1개의 이미지 URL로 불러오기(filePath: 폴더명/파일명)
     @GetMapping("/image")
-    public ResponseEntity<String> getImageUrl(@RequestParam String filePath) {
+    public ResponseEntity<String> getImageUrl(@RequestParam String filePath,
+                                              @CookieValue Long id) {
         try {
-            // Presigned URL 생성
-            String presignedUrl = s3ImageService.getPresignedUrl(filePath);
-            System.out.println("Presigned URL: " + presignedUrl);
-            return ResponseEntity.ok(presignedUrl);
+            if (signUpService.isUserActive(id)) {   // ACTIVE 유저 인 경우만
+                // Presigned URL 생성
+                String presignedUrl = s3ImageService.getPresignedUrl(filePath);
+                System.out.println("Presigned URL: " + presignedUrl);
+                return ResponseEntity.ok(presignedUrl);
+            } else {
+                return ResponseEntity.notFound().build();
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.badRequest().body(e.getMessage());
@@ -67,10 +84,14 @@ public class S3ImgController {
 
     // S3 버킷에 있는 이미지 URL로 불러오기
     @GetMapping("/images/urls")
-    public ResponseEntity<?> getAllImageUrls() {
+    public ResponseEntity<?> getAllImageUrls(@CookieValue Long id) {
         try {
-            Map<String, String> imageUrls = s3ImageService.getAllPresignedImageUrls();
-            return ResponseEntity.ok(imageUrls);
+            if (signUpService.isUserActive(id)) {   // ACTIVE 유저 인 경우만
+                Map<String, String> imageUrls = s3ImageService.getAllPresignedImageUrls();
+                return ResponseEntity.ok(imageUrls);
+            } else {
+                return ResponseEntity.notFound().build();
+            }
         } catch (Exception e) {
             log.error("S3 이미지 URL 전체 조회 중 오류 발생", e);
             return ResponseEntity.status(500).body("S3 이미지 URL 전체 조회 실패");
